Internalise component
