+++ src/work-allocation-2/components/task-list-filter/task-list-filter.component.html
@@ -1,0 +1,24 @@
+<div class="button-heading">
+  <button class="govuk-button hmcts-button--secondary" type="button" (click)="this.toggleFilter=!this.toggleFilter">
+    <ng-container *ngIf="!toggleFilter">
+      Show work filter
+    </ng-container>
+    <ng-container *ngIf="toggleFilter">
+      Hide work filter
+    </ng-container>
+  </button>
+  <ng-container *ngIf="selectedLocations.length && showFilteredText">
+    <span class="hmcts-badge hmcts-badge--black">Filtered view</span>
+    <span class="govuk-hint">All of your work may not be visible.</span>
+  </ng-container>
+</div>
+<div class="govuk-!-padding-3 filter govuk-!-margin-bottom-3" *ngIf="toggleFilter">
+  <h3 class="govuk-heading-m">{{jurisdiction}}</h3>
+  <xuilib-generic-filter
+    [config]="fieldsConfig"
+    [settings]="fieldsSettings"
+  ></xuilib-generic-filter>
+</div>
+<div class="govuk-!-padding-3 govuk-!-margin-bottom-3" *ngIf="selectedLocations.length">
+    Selected locations {{selectedLocations.length}}
+</div>
