+++ src/work-allocation-2/components/task-list-filter/task-list-filter.component.scss
@@ -1,0 +1,19 @@
+.filter {
+  background-color: #f2f2f1;
+  width: 75%;
+}
+
+.button-heading {
+  display: flex;
+  align-items: center;
+
+  span {
+    margin-bottom: 32px;
+    margin-left: 12px;
+  }
+
+  .hmcts-badge--black {
+    border-color: #0b0c0c;
+    color: #0b0c0c;
+  }
+}
