+++ src/work-allocation-2/components/task-list-filter/task-list-filter.component.spec.ts
@@ -1,28 +1,26 @@
 import { CdkTableModule } from '@angular/cdk/table';
-import { Component, ViewChild } from '@angular/core';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { Component, DebugElement, ViewChild } from '@angular/core';
+import { ComponentFixture, fakeAsync, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
 import { Router } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ExuiCommonLibModule, FilterService } from '@hmcts/rpx-xui-common-lib';
 import { of } from 'rxjs/internal/observable/of';
 
 import { ALL_LOCATIONS } from '../../../../api/workAllocation2/constants/locations';
-import { ErrorMessageComponent } from '../../../app/components';
-import { WorkAllocationComponentsModule } from '../../components/work-allocation.components.module';
 import { LocationDataService, WorkAllocationTaskService } from '../../services';
-import { InfoMessageContainerComponent } from '../info-message-container/info-message-container.component';
-import { TaskHomeComponent } from './task-home.component';
+import { TaskListFilterComponent } from './task-list-filter.component';
 
 @Component({
   template: `
-    <exui-task-home></exui-task-home>`
+    <exui-task-list-filter></exui-task-list-filter>`
 })
 class WrapperComponent {
-  @ViewChild(TaskHomeComponent) public appComponentRef: TaskHomeComponent;
+  @ViewChild(TaskListFilterComponent) public appComponentRef: TaskListFilterComponent;
 }
 
-describe('TaskHomeComponent', () => {
-  let component: TaskHomeComponent;
+describe('TaskListFilterComponent', () => {
+  let component: TaskListFilterComponent;
   let wrapper: WrapperComponent;
   let fixture: ComponentFixture<WrapperComponent>;
   let router: Router;
@@ -44,10 +42,9 @@ describe('TaskHomeComponent', () => {
         CdkTableModule,
         ExuiCommonLibModule,
         RouterTestingModule,
-        WorkAllocationComponentsModule,
         ExuiCommonLibModule
       ],
-      declarations: [TaskHomeComponent, WrapperComponent, InfoMessageContainerComponent, ErrorMessageComponent],
+      declarations: [TaskListFilterComponent, WrapperComponent ],
       providers: [
         { provide: WorkAllocationTaskService, useValue: mockTaskService },
         { provide: LocationDataService, useValue: { getLocations: () => of(ALL_LOCATIONS) } },
@@ -67,5 +64,29 @@ describe('TaskHomeComponent', () => {
   it('should create', () => {
     expect(component).toBeDefined();
   });
+
+  it('should show the toggle filter button', () => {
+    const button: DebugElement = fixture.debugElement.query(By.css('.govuk-button.hmcts-button--secondary'));
+    expect(button.nativeElement.innerText).toContain('Show work filter');
+  });
+
+  it('should select two locations', fakeAsync(() => {
+    const button: DebugElement = fixture.debugElement.query(By.css('.govuk-button.hmcts-button--secondary'));
+    button.nativeElement.click();
+
+    fixture.detectChanges();
+    const checkBoxes: DebugElement = fixture.debugElement.query(By.css('.govuk-checkboxes'));
+    const firstLocation = checkBoxes.nativeElement.children[0];
+    const secondLocation = checkBoxes.nativeElement.children[1];
+
+    firstLocation.click();
+    secondLocation.click();
+
+    fixture.detectChanges();
+    const applyButton: DebugElement = fixture.debugElement.query(By.css('#applyFilter'));
+    applyButton.nativeElement.click();
+    expect(component.selectedLocations.length).toEqual(2);
+
+  }));
 
 });
