+++ src/work-allocation-2/containers/task-home/task-home.component.ts
@@ -1,49 +1,21 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { ActivatedRoute, ActivatedRouteSnapshot, NavigationEnd, Router, RoutesRecognized } from '@angular/router';
-import { FilterService, FilterSetting, SubNavigation } from '@hmcts/rpx-xui-common-lib';
-import { FilterConfig, FilterFieldConfig } from '@hmcts/rpx-xui-common-lib/lib/models/filter.model';
+import { ActivatedRouteSnapshot, NavigationEnd, Router, RoutesRecognized } from '@angular/router';
+import { SubNavigation } from '@hmcts/rpx-xui-common-lib';
 import { Subscription } from 'rxjs';
-import { filter } from 'rxjs/operators';
 
-import { Location } from '../../models/dtos';
-import { TaskSortField } from '../../models/tasks';
-import { LocationDataService } from '../../services';
 import { AppUtils } from '../../../app/app-utils';
 import { ErrorMessage } from '../../../app/models';
+import { TaskSortField } from '../../models/tasks';
 
-export const LOCATION_ERROR: ErrorMessage = {
-  title: 'There is a problem',
-  description: 'At least one location is required',
-  fieldId: 'task_assignment_caseworker'
-};
 @Component({
   selector: 'exui-task-home',
   templateUrl: 'task-home.component.html',
   styleUrls: ['task-home.component.scss']
 })
 export class TaskHomeComponent implements OnInit, OnDestroy {
-  private static readonly FILTER_NAME = 'locations';
-  public toggleFilter = false;
-  public showFilteredText = false;
   public sortedBy: TaskSortField;
   public pageTitle: string;
-  public error: ErrorMessage;
-  public fieldsConfig: FilterConfig = {
-    persistence: 'local',
-    id: TaskHomeComponent.FILTER_NAME,
-    fields: [],
-    cancelButtonText: 'Reset to default',
-    applyButtonText: 'Apply',
-    cancelSetting: null
-  };
-  public allLocations: string[] = [];
-  public defaultLocations: string[] = [];
-  public locationFields: FilterSetting;
-  public fieldsSettings: FilterSetting = {
-    fields: [],
-    id: TaskHomeComponent.FILTER_NAME,
-  };
-  public selectedLocations: string[] = [];
+  public error: ErrorMessage = null;
   public jurisdiction: string = 'Immigration & Asylum';
   /**
    * Take in the Router so we can navigate when actions are clicked and
@@ -60,15 +32,9 @@ export class TaskHomeComponent implements OnInit, OnDestroy {
   ];
 
   private routeSubscription: Subscription;
-  private locationSubscription: Subscription;
-  private selectedLocationsSubscription: Subscription;
-  public errorSubscription: Subscription;
 
   constructor(
-    private readonly router: Router,
-    private readonly route: ActivatedRoute,
-    private readonly filterService: FilterService,
-    private readonly locationService: LocationDataService
+    private readonly router: Router
   ) {
   }
 
@@ -88,65 +54,11 @@ export class TaskHomeComponent implements OnInit, OnDestroy {
 
     // Set up the page data.
     this.setupPageData(this.router.routerState.root.snapshot);
-    this.locationSubscription = this.locationService.getLocations()
-      .subscribe((locations: Location[]) => {
-        locations.forEach((location) => this.allLocations.push(location.id.toString()));
-        this.setUpLocationFilter(locations);
-      });
-    this.errorSubscription = this.filterService.givenErrors.subscribe(value => {
-      if (value) {
-        this.error = LOCATION_ERROR;
-        this.error.description = value;
-      } else {
-        this.error = null;
-      }});
-    this.subscribeToSelectedLocations();
-    this.toggleFilter = false;
-  }
-
-  public resetToggle(): void {
-    this.toggleFilter = !this.toggleFilter;
-  }
-
-  // if there is no local storage available, default locations need to be reset
-  public getDefaultLocations(): string[] {
-    if (this.fieldsConfig && this.fieldsConfig.cancelSetting) {
-      this.fieldsConfig.cancelSetting.fields.forEach(field => {
-        if (field.name === 'locations') {
-          this.defaultLocations = field.value;
-        }
-      })
-    }
-    return this.defaultLocations;
-  }
-
-  public subscribeToSelectedLocations(): void {
-    this.selectedLocationsSubscription = this.filterService.getStream(TaskHomeComponent.FILTER_NAME)
-      .pipe(
-        filter((f: FilterSetting) => f && f.hasOwnProperty('fields'))
-      )
-      .subscribe((f: FilterSetting) => {
-        this.selectedLocations = f.fields.find((field) => field.name === TaskHomeComponent.FILTER_NAME).value;
-        const isFiltered = this.hasBeenFiltered(f, this.getDefaultLocations());
-        this.showFilteredText = isFiltered;
-        this.toggleFilter = false;
-      });
   }
 
   public ngOnDestroy(): void {
     if (this.routeSubscription) {
       this.routeSubscription.unsubscribe();
-    }
-    if (this.locationSubscription) {
-      this.locationSubscription.unsubscribe();
-    }
-
-    if (this.selectedLocationsSubscription) {
-      this.selectedLocationsSubscription.unsubscribe();
-    }
-
-    if (this.errorSubscription) {
-      this.errorSubscription.unsubscribe();
     }
   }
 
@@ -178,41 +90,7 @@ export class TaskHomeComponent implements OnInit, OnDestroy {
     return false;
   }
 
-  private hasBeenFiltered(f: FilterSetting, defaultLocations: string[]): boolean {
-    const selectedFields = f.fields.find(field => field.name === TaskHomeComponent.FILTER_NAME);
-    // check if selected fields are the same as cancelled filter settings
-    const containsNonDefaultFields = selectedFields.value.filter((v: string) => defaultLocations.indexOf(v) === -1).length > 0;
-    // check if the amount of fields selected is the same as the amount in the cancel settings
-    const notSameSize = !(defaultLocations.length === selectedFields.value.length);
-    return (containsNonDefaultFields || notSameSize) && (defaultLocations.length !== 0);
-  }
-
-  private setUpLocationFilter(locations: Location[]): void {
-    const field: FilterFieldConfig = {
-      name: TaskHomeComponent.FILTER_NAME,
-      options: locations.map((location) => ({
-        key: location.id,
-        label: location.locationName
-      })),
-      minSelected: 1,
-      maxSelected: 10,
-      minSelectedError: 'At least one location is required',
-      maxSelectedError: 'Maximum locations selected',
-      subTitle: 'Shows tasks and cases for the selected locations:',
-      type: 'checkbox'
-    };
-    if (this.route.snapshot.data && this.route.snapshot.data.location) {
-      const location: Location = this.route.snapshot.data.location;
-      this.defaultLocations = [`${location.id}`];
-    } else {
-      // as some judicial workers do not have a set location set their default to be all locations
-      this.defaultLocations = this.allLocations;
-    }
-    this.fieldsSettings.fields = [...this.fieldsSettings.fields, {
-      name: TaskHomeComponent.FILTER_NAME,
-      value: this.defaultLocations
-    }];
-    this.fieldsConfig.cancelSetting = JSON.parse(JSON.stringify(this.fieldsSettings));
-    this.fieldsConfig.fields.push(field);
+  public errorChangedHandler(error: ErrorMessage) {
+    this.error = error;
   }
 }
